requirements.txt
aider-chat==0.14.0
aiobotocore==2.5.0
aiofiles==22.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.18.0
alabaster==0.7.13
alembic==1.11.1
altair==5.0.1
ansiterm==1.1
anyio==3.6.2
apache-airflow==2.6.1
apache-airflow-providers-common-sql==1.5.1
apache-airflow-providers-ftp==3.4.1
apache-airflow-providers-http==4.4.1
apache-airflow-providers-imap==3.2.1
apache-airflow-providers-sqlite==3.4.1
apispec==5.2.2
apturl==0.5.2
argcomplete==3.0.8
argilla==1.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.4
botocore==1.29.76
Brlapi==0.8.3
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickclick==20.10.2
clickhouse-connect==0.5.20
cmake==3.26.4
colorama==0.4.4
colorlog==4.8.0
comm==0.1.2
command-not-found==0.3
commonmark==0.9.1
ConfigArgParse==1.7
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.1.0
cron-descriptor==1.4.0
croniter==1.3.15
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.5.7
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
diskcache==5.6.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.3.0
docutils==0.16
duckdb==0.7.1
duplicity==0.8.21
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.0
fasteners==0.14.1
fastjsonschema==2.16.3
filelock==3.11.0
Flask==2.2.5
Flask-AppBuilder==4.3.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
furo==2023.3.27
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
google-search-results==2.4.2
graph-tools==1.11
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
idna==3.4
imagesize==1.4.1
importlib-metadata==4.6.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonlines==4.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-cache==0.6.1
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
keyring==23.5.0
kiwisolver==1.4.4
langchain==0.0.136
langchainplus-sdk==0.0.17
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
limits==3.5.0
line-profiler==4.0.3
linkify-it-py==2.0.2
lit==16.0.6
livereload==2.6.3
llama-index==0.6.23
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
lz4==4.3.2
m2r2==0.3.2
macaroonbakery==1.3.1
Mako==1.1.3
mando==0.7.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
memory-profiler==0.61.0
mistune==0.8.4
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
myst-nb==0.17.2
myst-parser==0.18.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
ordered-set==4.1.0
package-name==0.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
perlcompat==1.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.0
plotly==5.15.0
pluggy==1.0.0
posthog==2.5.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyan3==1.2.0
pyarrow==12.0.0
pycairo==1.20.1
pycallgraph==1.1.3
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
pydeck==0.8.1b0
pyds @ git+https://github.com/jashandeep-sohi/pyds.git@ac022f87a2783a6e12fd6be386328b234c63a3bd
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pypandoc==1.11
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
Pytess==1.0.0
pytest==7.3.1
python-apt==2.4.0+ubuntu2
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-nvd3==0.15.0
python-pptx==0.6.21
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pyxdg==0.27
PyYAML==6.0
pyzmq==25.0.0
radon==6.0.1
rank-bm25==0.2.2
regex==2023.3.23
reportlab==3.6.8
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.5
rich_argparse==1.1.0
roman==3.3
s3fs==2023.5.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
snakeviz==2.2.0
sniffio==1.3.0
snowballstemmer==2.2.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinx-basic-ng==1.0.0b1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.26.1
streamlit==1.24.0
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tbdump==1.4
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
triton==2.0.0
tweepy==4.14.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
unstructured==0.5.11
uri-template==1.2.0
urllib3==2.0.2
usb-creator==0.3.7
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
xdg==5
xkit==0.0.0
XlsxWriter==3.0.9
xmltodict==0.13.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==1.0.0
zstandard==0.20.0
